% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_transformation_matrices.R
\name{bi_rotation_matrix}
\alias{bi_rotation_matrix}
\title{2D Rotation Matrix}
\usage{
bi_rotation_matrix(theta)
}
\arguments{
\item{theta}{numeric, rotation angle}
}
\value{
matrix 3x3
}
\description{
2D Rotation Matrix
}
\examples{
bi_rotation_matrix(pi/4)
}
