% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_transformation_matrices.R
\name{bi_scale_matrix}
\alias{bi_scale_matrix}
\title{2D Scaling Matrix}
\usage{
bi_scale_matrix(sx, sy)
}
\arguments{
\item{sx}{numeric, X-scale}

\item{sy}{numeric Y-scale}
}
\value{
matrix 3x3
}
\description{
2D Scaling Matrix
}
\examples{
bi_scale_matrix(2, 0.5)
}
