% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_transformation_matrices.R
\name{bi_shear_matrix}
\alias{bi_shear_matrix}
\title{2D Shearing Matrix}
\usage{
bi_shear_matrix(shx, shy)
}
\arguments{
\item{shx}{numeric, x-shear}

\item{shy}{numeric, y-shear}
}
\value{
matrix 3x3
}
\description{
2D Shearing Matrix
}
\examples{
bi_shear_matrix(0, 2)
}
