% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_transformation_matrices.R
\name{bi_tilt_matrix}
\alias{bi_tilt_matrix}
\title{2D Tilt Matrix}
\usage{
bi_tilt_matrix(e, f)
}
\arguments{
\item{e}{tilt parameter}

\item{f}{tilt parameter}
}
\value{
matrix 3x3
}
\description{
2D Tilt Matrix
}
\examples{
bi_tilt_matrix(0, 1)
}
