% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_matrix.R
\name{bi_transformation_matrix}
\alias{bi_transformation_matrix}
\title{2D transformation matrix, based on transformation type}
\usage{
bi_transformation_matrix(params, transformation)
}
\arguments{
\item{params}{a single row data.frame with parameters. Vector parameters must be name
with .X dimension suffix. E.g., scale.1 and scale.2.}

\item{transformation}{string, \code{"euclidean"}, \code{"affine"}, or \code{"projective"}}
}
\value{
matrix 3x3
}
\description{
2D transformation matrix, based on transformation type
}
\examples{
bi_transformation_matrix(
  data.frame(rotation=0,
             scale.1=1, scale.2=2,
             translation.1=0, translation.2=-5),
  "euclidean")
}
