% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_transformation_matrices.R
\name{bi_translation_matrix}
\alias{bi_translation_matrix}
\title{2D Translation Matrix}
\usage{
bi_translation_matrix(dx, dy)
}
\arguments{
\item{dx}{numeric, horizontal component}

\item{dy}{numeric, vertical component}
}
\value{
matrix 3x3
}
\description{
2D Translation Matrix
}
\examples{
bi_translation_matrix(1, -1)
}
