% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variables.R
\name{check_variables}
\alias{check_variables}
\title{Checks validity of variables' matrix}
\usage{
check_variables(var, var_label)
}
\arguments{
\item{var}{Matrix\link{N, ncol}}

\item{var_label}{Variable label for error messages}

\item{ncol}{Expected number of columns}
}
\description{
Checks whether the matrix is numeric,
has expected number of columns (\code{ncol}),
and has no missing/infinite data.
}
\examples{
# This should produce no error messages
check_variables(matrix(c(1, 2, 3, 4), ncol=2), "test matrix")
# This should produce non-numeric data error message
check_variables(matrix(as.character(c(1, 2, 3, 4)), ncol=2), "test matrix")
# These should produce non-finite data error message
check_variables(matrix(c(1, 2, NA, 4), ncol=2), "test matrix")
check_variables(matrix(c(1, 2, Inf, 4), ncol=2), "test matrix")
}
\keyword{internal}
