% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.tridim_transform}
\alias{coef.tridim_transform}
\title{Posterior distributions for transformation coefficients
in full or summarized form.}
\usage{
\method{coef}{tridim_transform}(object, summary = TRUE, probs = c(0.055, 0.945))
}
\arguments{
\item{object}{An object of class "tridim_transform".}

\item{summary}{Whether summary statistics should be returned instead of
raw sample values. Defaults to \code{TRUE}}

\item{probs}{The percentiles used to compute summary, defaults to 89\% credible inteval.}
}
\value{
If summary=FALSE, a list with matrix iterationsN x dimensionsN for
each variable.
If summary=TRUE, a data.frame with columns "dv{index}" with mean for each dependent
variable plus optional quantiles columns with names "dv{index}_{quantile}".
}
\description{
Posterior distributions for transformation coefficients
in full or summarized form.
}
\examples{
\dontrun{
euc2 <- fit_geometric_transformation(depV1+depV2~indepV1+indepV2,
                                        data = NakayaData,
                                        transformation = 'euclidean')

# full posterior distribution
transform_posterior <- coef(euc2, summary=FALSE)

# coefficients' summary with 89\% CI
coef(euc2)
}
}
