% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{female_face_neutral}
\alias{female_face_neutral}
\title{Female face, neutral expression}
\format{
A data frame with 64 observations on the following 3 variables:
\describe{
\item{x, y, z}{numeric vectors, 3D coordinates of factial landmarks.}
}
}
\usage{
female_face_neutral
}
\description{
Facial landmarks for a female face with neutral expression.
}
\keyword{datasets}
