% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_geometric_transformation.R
\name{fit_geometric_transformation}
\alias{fit_geometric_transformation}
\alias{fit_geometric_transformation.default}
\alias{fit_geometric_transformation.formula}
\title{Fitting Bidimensional or Tridimensional Regression / Geometric Transformation Models.}
\value{
returns an object of class \code{\link[TriDimRegression:tridim_transform-class]{tridim_transform-class}}.
}
\description{
Fits Bidimensional or Tridimensional regression / geometric transformation models using
Stan engine. Can be used either via a formula interface or by supplying two tables, one
for an indepdent variable (\code{iv}) and one for the dependent one (\code{dv}).
In the latter case, the two tables must have the same dimenisons (both N x 2 or N x 3).
}
\examples{
\dontrun{
# Geometric transformations of 2D data
euc2 <- fit_geometric_transformation(depV1 + depV2 ~ indepV1 + indepV2,
                                     NakayaData, 'euclidean')
aff2 <- fit_geometric_transformation(depV1 + depV2 ~ indepV1 + indepV2,
                                     NakayaData, 'affine')
prj2 <- fit_geometric_transformation(depV1 + depV2 ~ indepV1 + indepV2,
                                     NakayaData, 'projective')

# summary of transformation coefficients
coef(euc2)

# statistical comparison via WAIC criterion
loo::loo_compare(waic(euc2), waic(aff2), waic(prj2))
}
}
