% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tridim_transform-class.R
\name{is.tridim_transform}
\alias{is.tridim_transform}
\title{Checks if argument is a \code{tridim_transform} object}
\usage{
is.tridim_transform(x)
}
\arguments{
\item{x}{An \R object}
}
\value{
Logical
}
\description{
Checks if argument is a \code{tridim_transform} object
}
