% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.tridim_transform}
\alias{loo.tridim_transform}
\title{Computes an efficient approximate leave-one-out
cross-validation via loo library. It can be used
for a model comparison via loo::loo_compare() function.}
\usage{
\method{loo}{tridim_transform}(object)
}
\arguments{
\item{object}{A tridim_transform object}
}
\value{
A named list, see loo::loo() for details.
}
\description{
Computes an efficient approximate leave-one-out
cross-validation via loo library. It can be used
for a model comparison via loo::loo_compare() function.
}
\examples{
\dontrun{
euc2 <- fit_geometric_transformation(depV1+depV2~indepV1+indepV2, NakayaData, transformation = 'euclidean')
aff2 <- fit_geometric_transformation(depV1+depV2~indepV1+indepV2, NakayaData, transformation = 'affine')
loo::loo_compare(loo(euc2), loo(aff2))
}
}
