% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_transformation_matrices.R
\name{m2_euclidean}
\alias{m2_euclidean}
\title{2D Euclidean}
\usage{
m2_euclidean(a, b)
}
\arguments{
\item{a}{numeric, translation}

\item{b}{numeric, all other coefficients}
}
\value{
matrix 3x3
}
\description{
2D Euclidean
}
\examples{
m2_euclidean(c(2, 3), c(1, 0.5))
}
\keyword{internal}
