% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_transformation_matrices.R
\name{m3_euclidean_x}
\alias{m3_euclidean_x}
\title{3D Euclidean, rotation about x}
\usage{
m3_euclidean_x(a, b)
}
\arguments{
\item{a}{numeric, translation}

\item{b}{numeric, all other coefficients}
}
\value{
matrix 4x4
}
\description{
3D Euclidean, rotation about x
}
\examples{
m3_euclidean_x(c(2, 3, 1), c(0.5, 0.2))
}
\keyword{internal}
