% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.tridim_transform}
\alias{predict.tridim_transform}
\title{Computes posterior samples for the posterior predictive distribution.}
\usage{
\method{predict}{tridim_transform}(object, newdata = NULL, summary = TRUE, probs = NULL)
}
\arguments{
\item{object}{An object of class "tridim_transform"}

\item{newdata}{An optional two column data frame with independent variables.
If omitted, the fitted values are used.}

\item{summary}{Whether summary statistics should be returned instead of
raw sample values. Defaults to \code{TRUE}}

\item{probs}{The percentiles used to compute summary, defaults to NULL (no CI).}
}
\value{
If summary=FALSE, a numeric matrix \link{iterationsN, observationsN, variablesN}.
If summary=TRUE, a data.frame with columns "dv{index}" with mean for each dependent
variable plus optional quantiles columns with names "dv{index}_{quantile}".
}
\description{
Predicted values based on the bi/tridimensionsal regressional model object.
}
\examples{
\dontrun{
euc2 <- fit_geometric_transformation(depV1+depV2~indepV1+indepV2, NakayaData, transformation = 'euclidean')

# prediction summary
predictions <- predict(euc2)

# full posterior prediction samples
prediction_samples <- predict(euc2, summary=FALSE)
}
}
\seealso{
\code{\link{fit_geometric_transformation}}
}
