% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tridim_transform}
\alias{print.tridim_transform}
\title{Prints out tridim_transform object}
\usage{
\method{print}{tridim_transform}(x, ...)
}
\arguments{
\item{x}{A \link[=tridim_transform-class]{tridim_transform} object}

\item{...}{Unused}
}
\description{
Prints out tridim_transform object
}
\examples{
\dontrun{
euc2 <- fit_geometric_transformation(depV1+depV2~indepV1+indepV2,
                                     data = NakayaData,
                                     transformation = 'euclidean')
euc2
}
}
