% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{R2}
\alias{R2}
\alias{R2.tridim_transform}
\title{Computes R-squared or adjusted R-squared}
\usage{
\method{R2}{tridim_transform}(object, adjust = TRUE, summary = TRUE, probs = c(0.055, 0.945))
}
\arguments{
\item{object}{An object of class "tridim_transform"}

\item{adjust}{Whether R-squared should be adjusted, defaults to \code{TRUE}}

\item{summary}{Whether summary statistics should be returned instead of
raw sample values. Defaults to \code{TRUE}}

\item{probs}{The percentiles used to compute summary, defaults to 89\% credible interval.}
}
\value{
vector of values or a data.frame with summary
}
\description{
Computes R-squared or adjusted R-squared
}
\examples{
\dontrun{
euc2 <- fit_geometric_transformation(depV1+depV2~indepV1+indepV2,
  NakayaData, transformation = 'euclidean')
R2(euc2)
}
}
