% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_transformation_matrices.R
\name{tri_rotation_y_matrix}
\alias{tri_rotation_y_matrix}
\title{3D Rotation around y-axis}
\usage{
tri_rotation_y_matrix(theta)
}
\arguments{
\item{theta}{Rotation angle}
}
\value{
matrix 4x4
}
\description{
3D Rotation around y-axis
}
\examples{
tri_rotation_y_matrix(pi/4)
}
