% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_transformation_matrices.R
\name{tri_scale_matrix}
\alias{tri_scale_matrix}
\title{3D scale matrix}
\usage{
tri_scale_matrix(sx, sy, sz)
}
\arguments{
\item{sx}{x scaling}

\item{sy}{y scaling}

\item{sz}{z scaling}
}
\value{
matrix 4x4
}
\description{
3D scale matrix
}
\examples{
tri_scale_matrix(1, 2, 0.5)
}
