% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_transformation_matrices.R
\name{tri_shear_x_matrix}
\alias{tri_shear_x_matrix}
\title{3D Shear for x-axis}
\usage{
tri_shear_x_matrix(shx_y, shx_z)
}
\arguments{
\item{shx_y}{Shear for y-axis}

\item{shx_z}{Shear for z-axis}
}
\value{
matrix 4x4
}
\description{
3D Shear for x-axis
}
\examples{
tri_shear_x_matrix(2, 0.5)
}
