% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_transformation_matrices.R
\name{tri_shear_y_matrix}
\alias{tri_shear_y_matrix}
\title{3D Shear for y-axis}
\usage{
tri_shear_y_matrix(shy_x, shy_z)
}
\arguments{
\item{shx_x}{Shear for x-axis}

\item{shx_z}{Shear for z-axis}
}
\value{
matrix 4x4
}
\description{
3D Shear for y-axis
}
\examples{
tri_shear_y_matrix(2, 0.5)
}
