% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_transformation_matrices.R
\name{tri_shear_z_matrix}
\alias{tri_shear_z_matrix}
\title{3D Shear for z-axis}
\usage{
tri_shear_z_matrix(shz_x, shz_y)
}
\arguments{
\item{shz_x}{Shear for x-axis}

\item{shz_y}{Shear for z-axis}
}
\value{
matrix 4x4
}
\description{
3D Shear for z-axis
}
\examples{
tri_shear_z_matrix(2, 0.5)
}
