% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_matrix.R
\name{tri_transformation_matrix}
\alias{tri_transformation_matrix}
\title{3D transformation matrix, based on transformation type}
\usage{
tri_transformation_matrix(params, transformation)
}
\arguments{
\item{params}{a single row data.frame with parameters. Vector parameters must be name
with .X dimension suffix. E.g., scale.1 and scale.2.}

\item{transformation}{string, \code{"euclidean_x"}, \code{"euclidean_y"},
\code{"euclidean_z"}, \code{"euclidean"}, \code{"affine"}, or \code{"projective"}}
}
\value{
matrix 4x4
}
\description{
3D transformation matrix, based on transformation type
}
\examples{
tri_transformation_matrix(data.frame(rotation=0,
                                     scale.1=1, scale.2=2, scale.3=1,
                                     translation.1=0, translation.2=-5, translation.3=-4),
                         "euclidean_z")
}
