% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_transformation_matrices.R
\name{tri_translation_matrix}
\alias{tri_translation_matrix}
\title{3D Translation}
\usage{
tri_translation_matrix(dx, dy, dz)
}
\arguments{
\item{dx}{x translation}

\item{dy}{y translation}

\item{dz}{z translation}
}
\value{
matrix 4x4
}
\description{
3D Translation
}
\examples{
tri_translation_matrix(0, 10, 1)
}
