% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tridim_transform-class.R
\docType{class}
\name{tridim_transform-class}
\alias{tridim_transform-class}
\alias{tridim_transform}
\title{Class \code{tridim_transform} of geometric transformations
fitted with the \code{\link[TriDimRegression:fit_geometric_transformation]{fit_geometric_transformation}} function.}
\description{
Geometric transformations fitted with the
\code{\link[TriDimRegression:fit_geometric_transformation]{fit_geometric_transformation}} function
represented as a \code{tridim_transform} object with information about transformation, data dimension,
call formula, and fitted \code{\link[rstan:stanfit]{stanfit}} object,
}
\details{
See \code{methods(class = "tridim_transform")} for an overview of available methods.
}
\section{Slots}{

\describe{
\item{\code{transformation}}{A \code{string} with the transformation name.}

\item{\code{formula}}{A \code{\link[Formula:formula]{formula}} object.}

\item{\code{Ndim}}{An \code{integer} with data dimension, either \code{2} or \code{3}.}

\item{\code{data}}{A \code{list} containing variables used for the \code{\link[rstan:sampling]{sampling}}.}

\item{\code{stanmodel}}{A \code{\link[rstan:stan_model]{stan_model}} used for sampling.}
}}

\seealso{
\code{\link{fit_geometric_transformation}}
}
