% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriDimRegression-package.R
\docType{package}
\name{TriDimRegression-package}
\alias{TriDimRegression-package}
\alias{TriDimRegression}
\title{The 'TriDimRegression' package.}
\description{
Fits 2D and 3D geometric transformations. Provides posterior via Stan.
Includes computation of LOO and WAIC information criteria, R-squared.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
\seealso{
\code{\link[TriDimRegression:fit_geometric_transformation]{fit_geometric_transformation}}
\link[=tridim_transform-class]{tridim_transform}
}
