% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tridim_transform}
\alias{plot.tridim_transform}
\title{Posterior interval plots for key parameters. Uses bayesplot::mcmc_intervals.}
\usage{
\method{plot}{tridim_transform}(object, ...)
}
\arguments{
\item{object}{A tridim_transform object}

\item{...}{Extra parameters to be passed to bayesplot::mcmc_intervals}
}
\description{
Posterior interval plots for key parameters. Uses bayesplot::mcmc_intervals.
}
\examples{
\dontrun{
euc2 <- fit_geometric_transformation(depV1+depV2~indepV1+indepV2,
                                        data = NakayaData,
                                        transformation = 'euclidean')
plot(euc2)
}
}
