% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.tridim_transform}
\alias{summary.tridim_transform}
\title{Summary for a tridim_transform object}
\usage{
\method{summary}{tridim_transform}(object, ...)
}
\arguments{
\item{object}{A \link[=tridim_transform-class]{tridim_transform} object}

\item{...}{Unused}
}
\description{
Summary for a tridim_transform object
}
\examples{
\dontrun{
euc2 <- fit_geometric_transformation(depV1+depV2~indepV1+indepV2,
                                     data = NakayaData,
                                     transformation = 'euclidean')
summary(euc2)
}
}
